configs,stdlib: added a private l1, l2 and shared l3 cache
This change adds a private l1 private l2 and shared l3 cache
hierarchy as a stdlib component.
