+++ disaggregated_memory/cachehierarchies/private_l1_private_l2_shared_l3_cache_hierarchy.py
@@ -24,24 +24,33 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from gem5.components.cachehierarchies.classic.private_l1_private_l2_cache_hierarchy import (
-    PrivateL1PrivateL2CacheHierarchy,
-)
 from gem5.components.cachehierarchies.classic.caches.l1dcache import L1DCache
 from gem5.components.cachehierarchies.classic.caches.l1icache import L1ICache
 from gem5.components.cachehierarchies.classic.caches.l2cache import L2Cache
 from gem5.components.cachehierarchies.classic.caches.mmu_cache import MMUCache
+from gem5.components.cachehierarchies.classic.private_l1_private_l2_cache_hierarchy import PrivateL1PrivateL2CacheHierarchy
 from gem5.components.boards.abstract_board import AbstractBoard
 from gem5.isas import ISA
-from m5.objects import L2XBar
+
+from m5.objects import (
+    Cache,
+    L2XBar,
+    BaseXBar,
+    SystemXBar,
+    BadAddr,
+    Port
+)
 
 from gem5.utils.override import overrides
 
-from cachehierarchies.private_l1_private_l2_shared_l3_cache_hierarchy import (
-    PrivateL1PrivateL2SharedL3CacheHierarchy)
 
-class ClassicPrivateL1PrivateL2SharedL3SstDMCache(
-        PrivateL1PrivateL2SharedL3CacheHierarchy):
+class PrivateL1PrivateL2SharedL3CacheHierarchy(
+        PrivateL1PrivateL2CacheHierarchy):
+    """
+    A cache setup where each core has a private L1 Data and Instruction Cache,
+    and a private L2 cache.
+    """
+
     def __init__(
         self,
         l1d_size: str,
@@ -49,23 +58,42 @@ class ClassicPrivateL1PrivateL2SharedL3SstDMCache(
         l2_size: str,
         l3_size: str,
         l3_assoc: int = 16
-    ):
+    ) -> None:
+        """
+        :param l1d_size: The size of the L1 Data Cache (e.g., "32kB").
+        :type l1d_size: str
+        :param  l1i_size: The size of the L1 Instruction Cache (e.g., "32kB").
+        :type l1i_size: str
+        :param l2_size: The size of the L2 Cache (e.g., "256kB").
+        :type l2_size: str
+        :param membus: The memory bus. This parameter is optional parameter and
+        will default to a 64 bit width SystemXBar is not specified.
+
+        :type membus: BaseXBar
+        """
         super().__init__(
             l1d_size=l1d_size,
             l1i_size=l1i_size,
-            l2_size=l2_size,
-            l3_size=l3_size,
-            l3_assoc=l3_assoc
+            l2_size=l2_size
         )
 
-    @overrides(PrivateL1PrivateL2SharedL3CacheHierarchy)
+        self._l3_size = l3_size
+        self._l3_assoc = l3_assoc
+        self._l3_tag_latency = 20
+        self._l3_data_latency = 20
+        self._l3_response_latency = 40
+        self._l3_mshrs = 32
+        self._l3_tgts_per_mshr = 12
+
+
+    @overrides(PrivateL1PrivateL2CacheHierarchy)
     def incorporate_cache(self, board: AbstractBoard) -> None:
 
         # Set up the system port for functional access from the simulator.
         board.connect_system_port(self.membus.cpu_side_ports)
 
-        for cntr in board.get_local_memory().get_memory_controllers():
-            cntr.port = self.membus.mem_side_ports
+        for _, port in board.get_memory().get_mem_ports():
+            self.membus.mem_side_ports = port
 
         self.l1icaches = [
             L1ICache(size=self._l1i_size)
@@ -132,87 +160,3 @@ class ClassicPrivateL1PrivateL2SharedL3SstDMCache(
         self.l3bus.mem_side_ports = self.l3cache.cpu_side
         self.membus.cpu_side_ports = self.l3cache.mem_side
 
-class ClassicPrivateL1PrivateL2SstDMCache(PrivateL1PrivateL2CacheHierarchy):
-    def __init__(
-        self,
-        l1d_size: str,
-        l1i_size: str,
-        l2_size: str,
-    ) -> None:
-        """We need a specific version of DM caches for the external memory,
-        which does not connect the remote memory controller ports directly.
-
-        :param l1d_size: The size of the L1 Data Cache (e.g., "32kB").
-        :type l1d_size: str
-        :param  l1i_size: The size of the L1 Instruction Cache (e.g., "32kB").
-        :type l1i_size: str
-        :param l2_size: The size of the L2 Cache (e.g., "256kB").
-        :type l2_size: str
-        :param membus: The memory bus. This parameter is optional parameter and
-        will default to a 64 bit width SystemXBar is not specified.
-        :type membus: BaseXBar
-        """
-        super().__init__(l1i_size, l1d_size, l2_size)
-
-    @overrides(PrivateL1PrivateL2CacheHierarchy)
-    def incorporate_cache(self, board: AbstractBoard) -> None:
-
-        # Set up the system port for functional access from the simulator.
-        board.connect_system_port(self.membus.cpu_side_ports)
-
-        for cntr in board.get_local_memory().get_memory_controllers():
-            cntr.port = self.membus.mem_side_ports
-
-        self.l1icaches = [
-            L1ICache(size=self._l1i_size)
-            for i in range(board.get_processor().get_num_cores())
-        ]
-        self.l1dcaches = [
-            L1DCache(size=self._l1d_size)
-            for i in range(board.get_processor().get_num_cores())
-        ]
-        self.l2buses = [
-            L2XBar() for i in range(board.get_processor().get_num_cores())
-        ]
-        self.l2caches = [
-            L2Cache(size=self._l2_size)
-            for i in range(board.get_processor().get_num_cores())
-        ]
-        # ITLB Page walk caches
-        self.iptw_caches = [
-            MMUCache(size="8KiB")
-            for _ in range(board.get_processor().get_num_cores())
-        ]
-        # DTLB Page walk caches
-        self.dptw_caches = [
-            MMUCache(size="8KiB")
-            for _ in range(board.get_processor().get_num_cores())
-        ]
-
-        if board.has_coherent_io():
-            self._setup_io_cache(board)
-
-        for i, cpu in enumerate(board.get_processor().get_cores()):
-
-            cpu.connect_icache(self.l1icaches[i].cpu_side)
-            cpu.connect_dcache(self.l1dcaches[i].cpu_side)
-
-            self.l1icaches[i].mem_side = self.l2buses[i].cpu_side_ports
-            self.l1dcaches[i].mem_side = self.l2buses[i].cpu_side_ports
-            self.iptw_caches[i].mem_side = self.l2buses[i].cpu_side_ports
-            self.dptw_caches[i].mem_side = self.l2buses[i].cpu_side_ports
-
-            self.l2buses[i].mem_side_ports = self.l2caches[i].cpu_side
-
-            self.membus.cpu_side_ports = self.l2caches[i].mem_side
-
-            cpu.connect_walker_ports(
-                self.iptw_caches[i].cpu_side, self.dptw_caches[i].cpu_side
-            )
-
-            if board.get_processor().get_isa() == ISA.X86:
-                int_req_port = self.membus.mem_side_ports
-                int_resp_port = self.membus.cpu_side_ports
-                cpu.connect_interrupt(int_req_port, int_resp_port)
-            else:
-                cpu.connect_interrupt()
